scala3 support, springWithStart
scala3 required move to magnolia1 which is backward incompatable, so
we need two versions of Animatable in js/src/main/scala-2 and scala-3

springWithStart operates on a Signal[(A, Option[A])], taking an
an optional Animatable from which it is visually spawning.  Useful
when changing an element's parent or duplicating elements.